first run on getting all JUnit tests running again
