+++ ArgoWirelineFormat/src/test/resources/testResponsePayload.json
@@ -1,1 +1,1 @@
-{"responseID":"urn:uuid:9c6bd8aa-bf9b-45aa-9348-24897d89e86f","probeID":"--PROBE ID--","services":[{"id":"--SERVICE ID--","contractId":"--CONTRACT ID--","serviceName":"Some Service Name","consumability":"MACHINE_CONSUMABLE","description":"some service description","contractDescription":"some contract description","ttl":"0","accessPoints":[{"label":"internal","ipAddress":"some ip","port":"some port","url":"some url","dataType":"basic 1","data":"some data"}]}]}
+{"responseID":"urn:uuid:9c6bd8aa-bf9b-45aa-9348-24897d89e86f","probeID":"--PROBE ID--","services":[{"id":"--SERVICE ID--","serviceContractId":"--CONTRACT ID--","serviceName":"Some Service Name","consumability":"MACHINE_CONSUMABLE","description":"some service description","contractDescription":"some contract description","ttl":"0","accessPoints":[{"label":"internal","ipAddress":"some ip","port":"some port","url":"some url","dataType":"basic 1","data":"some data"}]}]}
