+++ ArgoWirelineFormat/src/test/resources/testResponsePayloadMultipleAccessPoints.json
@@ -1,1 +1,1 @@
-{"responseID":"urn:uuid:9c6bd8aa-bf9b-45aa-9348-24897d89e86f","probeID":"--PROBE ID--","services":[{"id":"--SERVICE ID--","contractId":"--CONTRACT ID--","serviceName":"Some Service Name","consumability":"MACHINE_CONSUMABLE","description":"some service description","contractDescription":"some contract description","accessPoints":[{"label":"internal","ipAddress":"some ip","port":"some port","url":"some url","dataType":"basic 1","data":"some data"},{"label":"external","ipAddress":"some ip 2","port":"some port 2","url":"some url 2","dataType":"basic 3","data":"some data 2"}]}]}
+{"responseID":"urn:uuid:9c6bd8aa-bf9b-45aa-9348-24897d89e86f","probeID":"--PROBE ID--","services":[{"id":"--SERVICE ID--","serviceContractId":"--CONTRACT ID--","serviceName":"Some Service Name","consumability":"MACHINE_CONSUMABLE","description":"some service description","contractDescription":"some contract description","accessPoints":[{"label":"internal","ipAddress":"some ip","port":"some port","url":"some url","dataType":"basic 1","data":"some data"},{"label":"external","ipAddress":"some ip 2","port":"some port 2","url":"some url 2","dataType":"basic 3","data":"some data 2"}]}]}
