+++ CLClient/src/main/java/ws/argo/CLClient/commands/util/Cache.java
@@ -11,7 +11,7 @@ import com.google.gson.JsonArray;
 import com.google.gson.JsonElement;
 import com.google.gson.JsonObject;
 
-import ws.argo.CLClient.listener.ResponseCache;
+import ws.argo.common.cache.ResponseCache;
 import ws.argo.wireline.response.JSONSerializer;
 import ws.argo.wireline.response.ResponseParseException;
 import ws.argo.wireline.response.ServiceWrapper;
