+++ CLClient/src/main/java/ws/argo/CLClient/listener/ProbeResponseResource.java
@@ -23,6 +23,8 @@ import javax.ws.rs.Path;
 import javax.ws.rs.Produces;
 
 import net.dharwin.common.tools.cli.api.console.Console;
+import ws.argo.common.cache.ExpiringService;
+import ws.argo.common.cache.ResponseCache;
 import ws.argo.wireline.response.JSONSerializer;
 import ws.argo.wireline.response.ResponseParseException;
 import ws.argo.wireline.response.ResponseWrapper;
