+++ CommonUtils/pom.xml
@@ -27,5 +27,10 @@
       <version>2.21</version>
     </dependency>
 
+    <dependency>
+      <groupId>ws.argo.wireline</groupId>
+      <artifactId>ArgoWirelineFormat</artifactId>
+      <version>0.4.1-SNAPSHOT</version>
+    </dependency>
   </dependencies>
 </project>
