+++ DemoWebClient/ArgoWebClient/.settings/org.eclipse.wst.common.component
@@ -4,10 +4,10 @@
         <wb-resource deploy-path="/WEB-INF/classes" source-path="/src/main/resources"/>
         <wb-resource deploy-path="/" source-path="/target/m2e-wtp/web-resources"/>
         <wb-resource deploy-path="/" source-path="/src/main/webapp" tag="defaultRootSource"/>
-        <dependent-module archiveName="ProbeSender-0.4.0-BETA.jar" deploy-path="/WEB-INF/lib" handle="module:/resource/ProbeSender/ProbeSender">
+        <dependent-module archiveName="ProbeSender-0.4.1-SNAPSHOT.jar" deploy-path="/WEB-INF/lib" handle="module:/resource/ProbeSender/ProbeSender">
             <dependency-type>uses</dependency-type>
         </dependent-module>
-        <dependent-module archiveName="ArgoWirelineFormat-0.4.0-BETA.jar" deploy-path="/WEB-INF/lib" handle="module:/resource/ArgoWirelineFormat/ArgoWirelineFormat">
+        <dependent-module archiveName="ArgoWirelineFormat-0.4.1-SNAPSHOT.jar" deploy-path="/WEB-INF/lib" handle="module:/resource/ArgoWirelineFormat/ArgoWirelineFormat">
             <dependency-type>uses</dependency-type>
         </dependent-module>
         <property name="java-output-path" value="/ArgoWebClient/target/classes"/>
