+++ DemoWebClient/AsynchListener/pom.xml
@@ -42,6 +42,11 @@
 			<artifactId>ArgoWirelineFormat</artifactId>
 			<version>0.4.1-SNAPSHOT</version>
 		</dependency>
+		<dependency>
+		  <groupId>ws.argo.common</groupId>
+		  <artifactId>CommonUtils</artifactId>
+		  <version>0.4.1-SNAPSHOT</version>
+		</dependency>
 	</dependencies>
 	<build>
 		<finalName>AsynchListener</finalName>
