+++ DemoWebClient/AsynchListener/src/main/java/ws/argo/DemoWebClient/AsynchListener/AsynchListener.java
@@ -22,8 +22,8 @@ import javax.ws.rs.POST;
 import javax.ws.rs.Path;
 import javax.ws.rs.Produces;
 
-import ws.argo.DemoWebClient.AsynchListener.ResponseCache.ExpiringService;
-import ws.argo.DemoWebClient.AsynchListener.ResponseCache.ResponseCache;
+import ws.argo.common.cache.ExpiringService;
+import ws.argo.common.cache.ResponseCache;
 import ws.argo.wireline.response.JSONSerializer;
 import ws.argo.wireline.response.ResponseParseException;
 import ws.argo.wireline.response.ResponseWrapper;
