+++ Responder/ResponderDaemon/src/test/java/ws/argo/responder/test/listener/ProbeResponseResource.java
@@ -22,6 +22,8 @@ import javax.ws.rs.POST;
 import javax.ws.rs.Path;
 import javax.ws.rs.Produces;
 
+import ws.argo.common.cache.ExpiringService;
+import ws.argo.common.cache.ResponseCache;
 import ws.argo.wireline.response.JSONSerializer;
 import ws.argo.wireline.response.ResponseParseException;
 import ws.argo.wireline.response.ResponseWrapper;
